Moves around files for new name
