--- bin/dynamodb-exporter
@@ -1,7 +1,0 @@
-#!/usr/bin/env ruby
-
-$: << File.absolute_path('lib')
-
-require 'dynamodb/exporter/app'
-
-Dynamodb::Exporter::App.start
