+++ bin/heckscher
@@ -1,0 +1,7 @@
+#!/usr/bin/env ruby
+
+$: << File.absolute_path('lib')
+
+require 'heckscher/app'
+
+Heckscher::App.start
