--- lib/dynamodb/exporter.rb
@@ -1,7 +1,0 @@
-require "dynamodb/exporter/version"
-require "dynamodb/exporter/export"
-
-module Dynamodb
-  module Exporter
-  end
-end
