--- lib/dynamodb/exporter/version.rb
@@ -1,5 +1,0 @@
-module Dynamodb
-  module Exporter
-    VERSION = "0.0.1"
-  end
-end
