+++ lib/heckscher.rb
@@ -1,0 +1,5 @@
+require "heckscher/version"
+require "heckscher/export"
+
+module Heckscher
+end
