+++ lib/heckscher/version.rb
@@ -1,0 +1,3 @@
+module Heckscher
+  VERSION = "0.0.1"
+end
